deeplearning/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
ultralytics==8.3.34
Pillow==9.5.0
numpy==1.24.0  # Changed from 1.23.5 for Python 3.12 compatibility
setuptools